requirements.txt
Flask>=3.0.2,<3.1
PyYAML==6.0.1
Markdown==3.5.2
python-dotenv>=1.0.0